requirements.txt
streamlit
google-genai
elevenlabs
sounddevice
numpy
scipy
python-dotenv
pyaudio
